LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/view-transition-types-mutable-no-document-element-crashtest-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'root.appendChild')
FAIL: Timed out waiting for notifyDone to be called
